refactor(widgets,shared): change of styles css, rename contacts
