+++ src/widgets/footer/ui/footer.tsx
@@ -2,11 +2,11 @@ import Link from "next/link";
 import { navLinks } from "@/shared/constants";
 import Image from "next/image";
 import { Button } from "@/shared/ui";
-import { Contacts } from "./constans";
+import { Contacts } from "./contacts";
 
 export function Footer() {
   return (
-    <footer className="grid justify-items-center gap-15 rounded-t-[20px] bg-accent-orange py-[50px] text-white xl:grid-cols-[auto_1fr] xl:grid-rows-1 xl:gap-15 xl:p-20">
+    <footer className="grid justify-items-center gap-15 rounded-t-[20px] bg-accent-orange py-[50px] text-white xl:grid-cols-[auto_1fr] xl:grid-rows-1 xl:p-20">
       <div className="flex flex-col items-center gap-8 xl:items-start">
         <Image
           src="/logo.png"
